This is a working version of the MyJavaDelegate Class. In the bpmn file a String variable with the name input_msg must contain a executable json. Space charactes in the json msg kann lead to errors.
A string variable called isSelfResetting can be used to automatically execute a reset after completion of the task.
If the status of the Task chances to Abborted, Abborting, Stopped, Stopping, Held or Holding an error event is activated. This error can lead to an alternative path in the bpmn diagram.